+++ R/support.export.R
@@ -166,9 +166,6 @@ brm.data <- function(formula, data = NULL, family = "gaussian", prior = list(),
       if (ncolZ[[i]] == 1) Z[[i]] <- as.vector(Z[[i]])
       for ( j in 1:length(name)) standata <- c(standata, setNames(list(eval(expr[j])), name[j]))
       if (g %in% names(cov.ranef)) {
-        if (length(r[[i]]) > 1) 
-          stop(paste("Currently, customized covariance structures are only implemented",
-                      "for grouping factors with a single random effect"))
         cov.ranef[[g]] <- as.matrix(cov.ranef[[g]])
         level.names <- rownames(cov.ranef[[g]])
         colnames(cov.ranef[[g]]) <- level.names
@@ -180,9 +177,10 @@ brm.data <- function(formula, data = NULL, family = "gaussian", prior = list(),
           stop(paste("Row names of covariance matrix of",g,"do not match names of the grouping levels"))
         if (!isSymmetric(unname(cov.ranef[[g]])))
           stop(paste("Covariance matrix of grouping factor",g,"is not symmetric"))
+        cov.ranef[[g]] <- cov.ranef[[g]][order(level.names), order(level.names)]
         cov.ranef[[g]] <- nrow(cov.ranef[[g]])/sum(diag(cov.ranef[[g]])) * cov.ranef[[g]]
-        cov.ranef[[g]] <- t(chol(cov.ranef[[g]][order(level.names), order(level.names)]))
-        standata <- c(standata, setNames(list(cov.ranef[[g]]), paste0("CFcov_",g)))
+        if (length(r[[i]]) == 1) cov.ranef[[g]] <- t(chol(cov.ranef[[g]]))
+        standata <- c(standata, setNames(list(cov.ranef[[g]]), paste0("cov_",g)))
       }
     }
   }